- pytest should work on any computer, just run pytest in terminal   - make sure requirements-dev.txt is installed `pip install requirements-dev.txt` - Frame objects now have repr - PEP8 refactoring - Fixed draw_bbox, servo_step and manual_test_cam_detect() bug

+++ camera.py
@@ -131,7 +131,7 @@ class Camera():
         """
 
         images: list[Frame] = []
-        servo_step = servo_range / num_images
+        servo_step = int(servo_range / num_images)
         MIN_ANGLE = int(c.config["CAMERASERVOS"]["min_angle"])
         MAX_ANGLE = int(c.config["CAMERASERVOS"]["max_angle"])
